Player page and items import (#15)
* Design how the user play page will work

* Update skill costs and XP

* Massive updates around the player in-game experience

* Got wounds and effects working

* Update gitignore for claude

* Lots of work on item import/export

+++ Threa.Dal/Dto/Character.cs
@@ -19,8 +19,8 @@ public class Character
     public string HairDescription { get; set; } = string.Empty;
     public string Description { get; set; } = string.Empty;
     public long Birthdate { get; set; }
-    public double XPTotal { get; set; }
-    public double XPBanked { get; set; }
+    public int XPTotal { get; set; }
+    public int XPBanked { get; set; }
     public int ActionPointMax { get; set; }
     public int ActionPointRecovery { get; set; }
     public int ActionPointAvailable { get; set; }
