+++ Threa.Dal/Dto/CharacterSkill.cs
@@ -3,5 +3,5 @@
 public class CharacterSkill : Skill
 {
     public int Level { get; set; }
-    public double XPBanked { get; set; }
+    public int XPBanked { get; set; }
 }
